[Refactor] AvailableRoomListScreen 관련 위젯 분리 #12
